requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-filter==24.2
psycopg2-binary==2.9.9
requests==2.31.0
stripe==8.7.0
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==22.0.0
drf-yasg~=1.21.10